refactor: refactor code of tasks create.js and copy.js
